[kernels] update size_t to uint32_t to reduce L1 GPU cache load

+++ include/kernels/curandom.cuh
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <curand_kernel.h>
 #include <cuda_runtime.h>
 
@@ -9,6 +10,6 @@ namespace tl::kernels {
     __global__ void init_curand_state(curandState* state, unsigned long long seed);
 
     // Weight and bias random number initializer kernel
-    __global__ void init_random_weights(curandState* state, float* buf, const size_t size);
+    __global__ void init_random_weights(curandState* state, float* buf, const uint32_t size);
 }
